算力字段类型修改

+++ user/user-server/src/main/java/com/mei/hui/user/entity/SysUser.java
@@ -90,7 +90,7 @@ public class SysUser {
     private Double totalBlockAward;
 
     @TableField(exist = false)
-    private Integer powerAvailable;
+    private Long powerAvailable;
 
     /** 角色对象 */
     @TableField(exist = false)
